+++ packages/vexide-devices/src/adi/gyroscope.rs
@@ -198,11 +198,9 @@ impl AdiGyroscope {
     }
 }
 
-impl AdiDevice for AdiGyroscope {
-    type PortNumberOutput = u8;
-
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+impl AdiDevice<1> for AdiGyroscope {
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {