+++ packages/vexide-devices/src/adi/light_sensor.rs
@@ -119,11 +119,9 @@ impl AdiLightSensor {
     }
 }
 
-impl AdiDevice for AdiLightSensor {
-    type PortNumberOutput = u8;
-
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+impl AdiDevice<1> for AdiLightSensor {
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {