+++ packages/vexide-devices/src/adi/line_tracker.rs
@@ -124,11 +124,9 @@ impl AdiLineTracker {
     }
 }
 
-impl AdiDevice for AdiLineTracker {
-    type PortNumberOutput = u8;
-
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+impl AdiDevice<1> for AdiLineTracker {
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {